визуальные улучшения

+++ app/src/main/java/ru/ssau/population/model/ChartState.kt
@@ -1,6 +1,6 @@
 package ru.ssau.population.model
 
 data class ChartState(
-    val t: List<Long>,        // моменты времени
-    val y: List<List<Long>>,  // список кривых (у каждой кривой список точек)
+    val t: List<Double>,        // моменты времени
+    val y: List<List<Double>>,  // список кривых (у каждой кривой список точек)
 )
